bug fixing for model run with reduced column set

+++ minos/minosPipeline/RunPipeline.py
@@ -276,9 +276,9 @@ def type_check(data):
 
     """
 
-    data['S7_mental_health'] = data['S7_mental_health'].astype(int)
-    data['S7_physical_health'] = data['S7_physical_health'].astype(int)
-    data['nutrition_quality_diff'] = data['nutrition_quality_diff'].astype(int)
+    #data['S7_mental_health'] = data['S7_mental_health'].astype(int)
+    #data['S7_physical_health'] = data['S7_physical_health'].astype(int)
+    #data['nutrition_quality_diff'] = data['nutrition_quality_diff'].astype(int)
     data['neighbourhood_safety'] = data['neighbourhood_safety'].astype(int)
     data['job_sec'] = data['job_sec'].astype(int)
     #data['S7_neighbourhood_safety'] = data['S7_neighbourhood_safety'].astype(str)
